fix(xfs): widen the bmbt child-loop counter to size_t
CodeQL flagged a narrow u16 loop counter compared against size_t maxrecs: an
extent fork with more than 65535 records (or a corrupt numrecs field) could
make u16 i wrap and walk the same children twice. The count is only bounded
by the fork buffer anyway, so it is widened to size_t.

+++ src/fs/xfs.cpp
@@ -394,7 +394,7 @@ ScanResult scan(DiskReader& disk, const ScanOptions& opt, Progress& prog) {
             u16 numrecs = f.be16(2);
             size_t maxrecs = (n.fork.size() - 4) / 16;
             std::vector<u64> children;
-            for (u16 i = 0; i < numrecs && i < maxrecs; i++) {
+            for (size_t i = 0; i < (size_t)numrecs && i < maxrecs; i++) {
                 size_t p = 4 + maxrecs * 8 + (size_t)i * 8;
                 if (!f.has(p, 8)) break;
                 children.push_back(f.be64(p));
